test(webchat): extract parameterized gateway fixture
